refactor(frontend): 🏗️ change frontend folder structure
This way files are more neatly organized
