+++ app/frontend/src/features/auth/hooks/use-auth.ts
@@ -1,4 +1,4 @@
-import { AuthContext } from "@/context/auth-provider"; // Adjust the import path as necessary
+import { AuthContext } from "@/features/auth/context/auth-provider"; // Adjust the import path as necessary
 import { useContext } from "react";
 
 export function useAuth() {
