+++ app/frontend/src/features/auth/queries.ts
