+++ app/frontend/src/features/chat/hooks/use-message-service.ts
@@ -1,5 +1,5 @@
-import { useAuth } from "@/hooks/auth";
-import { MessageService } from "@/lib/message-service";
+import { useAuth } from "@/features/auth/hooks/use-auth";
+import { MessageService } from "@/features/chat/message-service";
 import { useCallback } from "react";
 
 // * Hook to use the message service
