+++ app/frontend/src/features/chat/message-service.ts
@@ -1,20 +1,4 @@
-import Dexie from "dexie";
-import type { EntityTable } from "dexie";
-
-export interface Message {
-	id?: string;
-	content: string;
-	timestamp: number;
-	userId: string;
-}
-
-const messageDb = new Dexie("MessageDatabase") as Dexie & {
-	messages: EntityTable<Message, "id">;
-};
-
-messageDb.version(5).stores({
-	messages: "++id, content, timestamp, userId",
-});
+import { type Message, messageDb } from "@/features/db";
 
 // * MessageService singleton for managing messages
 class MessageService {
